Update 7-9-2025 Cleaned up function docstrings.
Added class option "enable_command_logging". Set to false by default. When enabled, logs queries and parameters being executed.

Fixed a bug in the load_config function that would cause the return statement to not be reached.

Added count function

Added class entry and exit functions so that SQLiteHelper may be used with "with" statements.